Satisfy clippy and some code re-organization
